refactor: [UF-103] StorageProgress 삭제 후 Progress로 통일 #77
