+++ src/components/ui/Progress/index.ts
@@ -1,3 +1,2 @@
 export { Progress } from './Progress';
-export { StorageProgress } from './StorageProgress';
 export type { ProgressProps } from './Progress.types';
